// web/admin/gql/schema.tsx
export const crs = { type: 'name', properties: { name: 'urn:ogc:def:crs:EPSG::4326' } }
export type Maybe<T> = T | null | undefined
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] }

// Generated on 2021-10-04T01:13:08+05:30


/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string
  String: string
  Boolean: boolean
  Int: number
  Float: number
  /** A location in a connection that can be used for resuming pagination. */
  Cursor: any
  /** A universally unique identifier as defined by [RFC 4122](https://tools.ietf.org/html/rfc4122). */
  UUID: string
  /** A point in time as described by the [ISO 8601](https://en.wikipedia.org/wiki/ISO_8601) standard. May or may not include a timezone. */
  Datetime: Date
  /** The `JSON` scalar type represents JSON values as specified by [ECMA-404](http://www.ecma-international.org/publications/files/ECMA-ST/ECMA-404.pdf). */
  JSON: { [key: string]: unknown }
  /** A floating point number that requires more precision than IEEE 754 binary 64 */
  BigFloat: any
  /** The `GeoJSON` scalar type represents GeoJSON values as specified by[RFC 7946](https://tools.ietf.org/html/rfc7946). */
  GeoJSON: GeometryObject
  Email: string
  /** Relaxed phone number validation */
  PhoneUs: any
  /** The day, does not include a time. */
  Date: any
  /** The exact time of day, does not include the date. May or may not have a timezone offset. */
  Time: string
  /** A JSON Web Token defined by [RFC 7519](https://tools.ietf.org/html/rfc7519) which securely represents claims between two parties. */
  JwtToken: string
}
// ...
export type Node = {
  /** A globally unique identifier. Can be used in various places throughout the system to identify this single value. */
  nodeId: Scalars['ID']
}
// ...
export type Amenity = Node & {
  /** A globally unique identifier. Can be used in various places throughout the system to identify this single value. */
  nodeId: Scalars['ID']
  id: Scalars['UUID']
  name: Scalars['String']
  weight?: Maybe<Scalars['Int']>
  description?: Maybe<Scalars['String']>
  status: ContentStatusT
  createdAt: Scalars['Datetime']
  updatedAt?: Maybe<Scalars['Datetime']>
  slug?: Maybe<Scalars['String']>
  /** Reads and enables pagination through a set of `SlotAmenity`. */
  slotAmenities: SlotAmenitiesConnection
  /** Reads and enables pagination through a set of `SlotAmenity`. */
  slotAmenitiesList: Array<SlotAmenity>
  /** Reads and enables pagination through a set of `Slot`. */
  slotsBySlotAmenityAmenityIdAndSlotId: AmenitySlotsBySlotAmenityAmenityIdAndSlotIdManyToManyConnection
  /** Reads and enables pagination through a set of `Slot`. */
  slotsBySlotAmenityAmenityIdAndSlotIdList: Array<Slot>
}
// ...
export type SlotAmenity = Node & {
  /** A globally unique identifier. Can be used in various places throughout the system to identify this single value. */
  nodeId: Scalars['ID']
  id: Scalars['UUID']
  slotId: Scalars['UUID']
  amenityId: Scalars['UUID']
  /** Reads a single `Slot` that is related to this `SlotAmenity`. */
  slot?: Maybe<Slot>
  /** Reads a single `Amenity` that is related to this `SlotAmenity`. */
  amenity?: Maybe<Amenity>
}

export type Slot = Node & {
  /** A globally unique identifier. Can be used in various places throughout the system to identify this single value. */
  nodeId: Scalars['ID']
  id: Scalars['UUID']
  name: Scalars['String']
  ownerId: Scalars['UUID']
  vehicleSizeId: Scalars['UUID']
  address?: Maybe<Scalars['JSON']>
  timezone: Scalars['String']
  pricePerHour: Scalars['BigFloat']
  status: SlotStatusT
  photoUrl?: Maybe<Scalars['String']>
  description?: Maybe<Scalars['String']>
  notes?: Maybe<Scalars['String']>
  slug?: Maybe<Scalars['String']>
  createdAt: Scalars['Datetime']
  updatedAt?: Maybe<Scalars['Datetime']>
  deleted: Scalars['Boolean']
  deletedAt?: Maybe<Scalars['Datetime']>
  verificationStatus: SlotVerificationStatus
  parkingSpaceId?: Maybe<Scalars['UUID']>
  location: GeometryPoint
  shape?: Maybe<GeometryPolygon>
  accessRestrictions?: Maybe<AccessRestrictions>
  businessStatus?: Maybe<SlotBusinessStatus>
  businessStatusReason?: Maybe<Scalars['String']>
  category?: Maybe<SlotCategory>
  contributorId?: Maybe<Scalars['UUID']>
  level?: Maybe<Scalars['Int']>
  mapSource?: Maybe<Scalars['UUID']>
  slotDimensions?: Maybe<Scalars['JSON']>
  tempUnavailable?: Maybe<Scalars['Boolean']>
  tempUnavailableFrom?: Maybe<Scalars['Datetime']>
  tempUnavailableTo?: Maybe<Scalars['Datetime']>
  waypoints?: Maybe<Scalars['JSON']>
  /** Reads a single `User` that is related to this `Slot`. */
  owner?: Maybe<User>
  /** Reads a single `VehicleSize` that is related to this `Slot`. */
  vehicleSize?: Maybe<VehicleSize>
  /** Reads a single `ParkingSpace` that is related to this `Slot`. */
  parkingSpace?: Maybe<ParkingSpace>
  /** Reads and enables pagination through a set of `SlotAmenity`. */
  slotAmenities: SlotAmenitiesConnection
  /** Reads and enables pagination through a set of `SlotAmenity`. */
  slotAmenitiesList: Array<SlotAmenity>
  /** Reads and enables pagination through a set of `SlotAvailability`. */
  slotAvailabilities: SlotAvailabilitiesConnection
  /** Reads and enables pagination through a set of `SlotAvailability`. */
  slotAvailabilitiesList: Array<SlotAvailability>
  /** Reads and enables pagination through a set of `SlotBooking`. */
  slotBookings: SlotBookingsConnection
  /** Reads and enables pagination through a set of `SlotBooking`. */
  slotBookingsList: Array<SlotBooking>
  /** Reads and enables pagination through a set of `Amenity`. */
  amenitiesBySlotAmenitySlotIdAndAmenityId: SlotAmenitiesBySlotAmenitySlotIdAndAmenityIdManyToManyConnection
  /** Reads and enables pagination through a set of `Amenity`. */
  amenitiesBySlotAmenitySlotIdAndAmenityIdList: Array<Amenity>
  /** Reads and enables pagination through a set of `Currency`. */
  currenciesBySlotAvailabilitySlotIdAndTariffCurrency: SlotCurrenciesBySlotAvailabilitySlotIdAndTariffCurrencyManyToManyConnection
  /** Reads and enables pagination through a set of `Currency`. */
  currenciesBySlotAvailabilitySlotIdAndTariffCurrencyList: Array<Currency>
  /** Reads and enables pagination through a set of `User`. */
  usersBySlotBookingSlotIdAndUserId: SlotUsersBySlotBookingSlotIdAndUserIdManyToManyConnection
  /** Reads and enables pagination through a set of `User`. */
  usersBySlotBookingSlotIdAndUserIdList: Array<User>
  /** Reads and enables pagination through a set of `PaymentReceipt`. */
  paymentReceiptsBySlotBookingSlotIdAndPaymentReceiptId: SlotPaymentReceiptsBySlotBookingSlotIdAndPaymentReceiptIdManyToManyConnection
  /** Reads and enables pagination through a set of `PaymentReceipt`. */
  paymentReceiptsBySlotBookingSlotIdAndPaymentReceiptIdList: Array<PaymentReceipt>
}
// ...
export enum SlotStatusT {
  Enabled = 'ENABLED',
  Disabled = 'DISABLED',
  Unlisted = 'UNLISTED',
  Deleted = 'DELETED',
}

/** Verification status for slots */
export enum SlotVerificationStatus {
  Pending = 'PENDING',
  Rejected = 'REJECTED',
  Verified = 'VERIFIED',
}

export type GeometryPoint = GeometryInterface &
  GeometryGeometry & {
    geojson?: Maybe<Scalars['GeoJSON']>
    srid: Scalars['Int']
    x: Scalars['Float']
    y: Scalars['Float']
  }
// ...
export type GeometryPolygon = GeometryInterface &
  GeometryGeometry & {
    geojson?: Maybe<Scalars['GeoJSON']>
    srid: Scalars['Int']
    exterior?: Maybe<GeometryLineString>
    interiors?: Maybe<Array<Maybe<GeometryLineString>>>
  }

export type GeometryLineString = GeometryInterface &
  GeometryGeometry & {
    geojson?: Maybe<Scalars['GeoJSON']>
    srid: Scalars['Int']
    points?: Maybe<Array<Maybe<GeometryPoint>>>
  }

/** Allowed values for access_restrictions are: none, barrier, keycard, keycode */
export enum AccessRestrictions {
  None = 'NONE',
  Barrier = 'BARRIER',
  Keycard = 'KEYCARD',
  Keycode = 'KEYCODE',
}

/** values for business_status are: active, inactive */
export enum SlotBusinessStatus {
  Active = 'ACTIVE',
  Inactive = 'INACTIVE',
}

/** Allowed values for slot_category are: private, public */
export enum SlotCategory {
  Private = 'PRIVATE',
  Public = 'PUBLIC',
}
// ...
export type VehicleSize = Node & {
  /** A globally unique identifier. Can be used in various places throughout the system to identify this single value. */
  nodeId: Scalars['ID']
  id: Scalars['UUID']
  status: ContentStatusT
  name?: Maybe<Scalars['String']>
  weight: Scalars['Int']
  description?: Maybe<Scalars['String']>
  createdAt: Scalars['Datetime']
  updatedAt?: Maybe<Scalars['Datetime']>
  /** Reads and enables pagination through a set of `Slot`. */
  slots: SlotsConnection
  /** Reads and enables pagination through a set of `Slot`. */
  slotsList: Array<Slot>
  /** Reads and enables pagination through a set of `Vehicle`. */
  vehicles: VehiclesConnection
  /** Reads and enables pagination through a set of `Vehicle`. */
  vehiclesList: Array<Vehicle>
  /** Reads and enables pagination through a set of `User`. */
  usersBySlotVehicleSizeIdAndOwnerId: VehicleSizeUsersBySlotVehicleSizeIdAndOwnerIdManyToManyConnection
  /** Reads and enables pagination through a set of `User`. */
  usersBySlotVehicleSizeIdAndOwnerIdList: Array<User>
  /** Reads and enables pagination through a set of `ParkingSpace`. */
  parkingSpacesBySlotVehicleSizeIdAndParkingSpaceId: VehicleSizeParkingSpacesBySlotVehicleSizeIdAndParkingSpaceIdManyToManyConnection
  /** Reads and enables pagination through a set of `ParkingSpace`. */
  parkingSpacesBySlotVehicleSizeIdAndParkingSpaceIdList: Array<ParkingSpace>
  /** Reads and enables pagination through a set of `GeodataProvider`. */
  geodataProvidersBySlotVehicleSizeIdAndMapSource: VehicleSizeGeodataProvidersBySlotVehicleSizeIdAndMapSourceManyToManyConnection
  /** Reads and enables pagination through a set of `GeodataProvider`. */
  geodataProvidersBySlotVehicleSizeIdAndMapSourceList: Array<GeodataProvider>
  /** Reads and enables pagination through a set of `User`. */
  usersByVehicleVehicleSizeIdAndOwnerId: VehicleSizeUsersByVehicleVehicleSizeIdAndOwnerIdManyToManyConnection
  /** Reads and enables pagination through a set of `User`. */
  usersByVehicleVehicleSizeIdAndOwnerIdList: Array<User>
  /** Reads and enables pagination through a set of `VehicleType`. */
  vehicleTypesByVehicleVehicleSizeIdAndVehicleTypeId: VehicleSizeVehicleTypesByVehicleVehicleSizeIdAndVehicleTypeIdManyToManyConnection
  /** Reads and enables pagination through a set of `VehicleType`. */
  vehicleTypesByVehicleVehicleSizeIdAndVehicleTypeIdList: Array<VehicleType>
}
// ...
export type CreateSlotMutation = {
  createSlot?: Maybe<{
    slot?: Maybe<{
      id: string
      name: string
      timezone: string
      slug?: Maybe<string>
      accessRestrictions?: Maybe<AccessRestrictions>
      parkingSpaceId?: Maybe<string>
      level?: Maybe<number>
      notes?: Maybe<string>
      address?: Maybe<{ [key: string]: unknown }>
      tempUnavailable?: Maybe<boolean>
      tempUnavailableFrom?: Maybe<Date>
      tempUnavailableTo?: Maybe<Date>
      location: { longitude: number; latitude: number }
      slotAmenitiesList: Array<{ amenity?: Maybe<{ id: string; name: string }> }>
      slotAvailabilitiesList: Array<{
        id: string
        dayOfWeek: number
        startHour: string
        endHour: string
      }>
      shape?: Maybe<{ geojson?: Maybe<GeometryObject> }>
    }>
  }>
}

export type DeleteSlotMutationVariables = Exact<{
  id: Scalars['UUID']
}>

export type DeleteSlotMutation = { deleteSlot?: Maybe<{ deletedSlotNodeId?: Maybe<string> }> }

export type UpdateSlotMutationVariables = Exact<{
  id: Scalars['UUID']
  patch: SlotPatch
}>

export type UpdateSlotMutation = {
  updateSlot?: Maybe<{ slot?: Maybe<{ id: string; name: string; status: SlotStatusT }> }>
}
// ...
export type ManageSlotQuery = {
  slot?: Maybe<{
    id: string
    name: string
    notes?: Maybe<string>
    timezone: string
    accessRestrictions?: Maybe<AccessRestrictions>
    category?: Maybe<SlotCategory>
    level?: Maybe<number>
    status: SlotStatusT
    description?: Maybe<string>
    pricePerHour: any
    photoUrl?: Maybe<string>
    address?: Maybe<{ [key: string]: unknown }>
    vehicleSizeId: string
    verificationStatus: SlotVerificationStatus
    parkingSpaceId?: Maybe<string>
    businessStatusReason?: Maybe<string>
    tempUnavailable?: Maybe<boolean>
    tempUnavailableFrom?: Maybe<Date>
    tempUnavailableTo?: Maybe<Date>
    waypoints?: Maybe<{ [key: string]: unknown }>
    slotAmenitiesList: Array<{ amenityId: string }>
    slotAvailabilitiesList: Array<{
      id: string
      dayOfWeek: number
      startHour: string
      endHour: string
    }>
    owner?: Maybe<{ id: string; name: string; email: string }>
    location: { longitude: number; latitude: number }
    shape?: Maybe<{ geojson?: Maybe<GeometryObject> }>
  }>
}